Add valitaion error handlers
